re US5459 re DE2990 re DE2991 Make the securityEvent resource implementation contained entirely in the SecurityEventResource and SecurityEventContextResolver classes
This should

* enforce the SECURITY_EVENT_VIEW right for all endpoints
* only permit the /api/rest/securityEvent and /api/rest/securityEvent/{id} endpoints

The securityEvent entity was added to the list of entities to ignore for the CommonValidatorTest, as it is no longer possible
to create them from the API by anyone, let alone staff users.

The SecurityEventResourceTest was modified to create security events directly in the repository ("database") rather than
through the creation method of the security event resource, as creating security events through the API is forbidden
for everyone.


Former-commit-id: 7abe75ee118f49f06ab1f1516bd9ae8ed9b0c426